Initial commit
So far I have been learning some PHP and how to use triggers with MySQL. I have built two forms for creating an email and creating a calculation. When either of these are created a trigger causes a notification of the appropriate type to be added to the notifications table. I am currently working out a way to display up to date notifications on the webpage.